requirements.txt
openai
praw
feedparser
requests
python-dotenv
faker
stem
playwright
flask
google-api-python-client
crewai
stripe
redis
docker
psutil
schedule
